Fix memory leak due to unclosed resources
A severe memory leak was found when letting KirBot run with 256MB of heap for 48 hours. 

At the end of the 48 hours, the robot was found to have run out of memory. Upon furhter investigation it appeared that the database connection was leaking as PreparedStatements and ResultSets were not being closed when they were done.

This was fixed by properly closing result sets and prepared statements.